product-dashboard-page:add skeleton loading to all categories page and creator folder for each page
